Save regions as image when saving the regions data
